feat: move player
