+++ packages/player/bun.build.ts
@@ -1,0 +1,6 @@
+import baseBuilder from '../../bun.build.base';
+
+await baseBuilder({
+  entrypoints: ['./src/player.ts', './src/hls-pipeline.ts', './src/dash-pipeline.ts'],
+  outdir: './dist',
+});
