+++ packages/player/jest.config.js
@@ -1,0 +1,11 @@
+// eslint-disable-next-line import/extensions
+import base from '../../jest.config.base.js';
+
+/** @type {import('ts-jest').JestConfigWithTsJest} */
+const config = {
+  ...base,
+  displayName: 'player',
+  coverageDirectory: './coverage',
+};
+
+export default config;
