+++ packages/player/package.json
@@ -1,6 +1,26 @@
 {
   "name": "@web-media-box/player",
   "version": "0.0.0",
-  "dependencies": {
+  "type": "module",
+  "scripts": {
+    "build": "bun run build:source && bun run build:types",
+    "build:types": "dts-bundle-generator --config .dts-bundle-generator.config.json",
+    "build:source": "bun run bun.build.ts",
+    "test": "bunx jest",
+    "test:coverage": "bunx jest --coverage=true"
+  },
+  "exports": {
+    "./player": {
+      "types": "./dist/player.d.ts",
+      "default": "./dist/player.js"
+    },
+    "./hls-pipeline": {
+      "types": "./dist/hls-pipeline.d.ts",
+      "default": "./dist/hls-pipeline.js"
+    },
+    "./dash-pipeline": {
+      "types": "./dist/dash-pipeline.d.ts",
+      "default": "./dist/dash-pipeline.js"
+    }
   }
 }
