+++ packages/player/src/dash-pipeline.ts
@@ -1,0 +1,1 @@
+export { DashPipeline } from './lib/pipelines/mse/dash';
