+++ packages/player/src/hls-pipeline.ts
@@ -1,0 +1,1 @@
+export { HlsPipeline } from './lib/pipelines/mse/hls';
