+++ packages/player/src/lib/configuration.ts
