+++ packages/player/src/lib/errors.ts
