+++ packages/player/src/lib/events.ts
@@ -1,4 +1,4 @@
-import type PlayerError from '@/player/errors.ts';
+import type PlayerError from './errors';
 
 export const Events = {
   EnterPictureInPictureMode: 'EnterPictureInPictureMode',
