+++ packages/player/src/lib/pipelines/mse/dash/index.ts
@@ -1,0 +1,1 @@
+export { default as DashPipeline } from './dashPipeline';
