+++ packages/player/src/lib/pipelines/mse/hls/index.ts
@@ -1,0 +1,1 @@
+export { default as HlsPipeline } from './hlsPipeline';
