+++ packages/player/src/lib/utils/eventEmitter.ts
