+++ packages/player/src/player.ts
@@ -1,0 +1,1 @@
+export { default as Player } from './lib/player';
