+++ packages/player/test/index.test.ts
@@ -1,4 +1,4 @@
-import { describe, it, expect } from 'bun:test';
+import { describe, it, expect } from '@jest/globals';
 
 describe('player spec', () => {
   it('mock spec', () => {
