--- packages/web-media-box/README.md
@@ -1,1 +1,0 @@
-🚧🚧🚧🚧🚧🚧🚧🚧🚧🚧🚧
