--- packages/web-media-box/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "web-media-box",
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "build": "bun run build:player & bun run build:dts",
-    "build:player": "bun build.js",
-    "build:dts": "dts-bundle-generator --config .dts-bundle-generator.config.json"
-  },
-  "exports": {
-    "./hls-parser": {
-      "types": "./dist/hls-parser/index.d.ts",
-      "default": "./dist/hls-parser/index.js"
-    },
-    "./dash-parser": {
-      "types": "./dist/dash-parser/index.d.ts",
-      "default": "./dist/dash-parser/index.js"
-    },
-    "./player": {
-      "types": "./dist/player/index.d.ts",
-      "default": "./dist/player/index.js"
-    },
-    "./pipelines/mse/hls": {
-      "types": "./dist/pipelines/mse/hls/index.d.ts",
-      "default": "./dist/pipelines/mse/hls/index.js"
-    },
-    "./pipelines/mse/dash": {
-      "types": "./dist/pipelines/mse/dash/index.d.ts",
-      "default": "./dist/pipelines/mse/dash/index.js"
-    },
-    "./pipelines/mse": {
-      "types": "./dist/pipelines/mse/index.d.ts",
-      "default": "./dist/pipelines/mse/index.js"
-    }
-  }
-}
