--- packages/web-media-box/src/pipelines/mse/dash/index.ts
@@ -1,1 +1,0 @@
-export { default as DashPipeline } from './dashPipeline.ts';
