--- packages/web-media-box/src/pipelines/mse/hls/index.ts
@@ -1,1 +1,0 @@
-export { default as HlsPipeline } from './hlsPipeline.ts';
