--- packages/web-media-box/src/player/index.ts
@@ -1,1 +1,0 @@
-export { default as Player } from './player.ts';
